Moving actions to correct files
